+++ tests/ecdsa/mod.rs
@@ -25,6 +25,7 @@ fn round_trip_sign_verification_pk8() {
     assert!(is_valid);
 }
 
+#[cfg(feature = "use_pem")]
 #[test]
 fn round_trip_sign_verification_pem() {
     let privkey_pem = include_bytes!("private_ecdsa_key.pem");
@@ -42,6 +43,7 @@ fn round_trip_sign_verification_pem() {
     assert!(is_valid);
 }
 
+#[cfg(feature = "use_pem")]
 #[test]
 fn round_trip_claim() {
     let privkey_pem = include_bytes!("private_ecdsa_key.pem");
@@ -67,6 +69,7 @@ fn round_trip_claim() {
 }
 
 // https://jwt.io/ is often used for examples so ensure their example works with jsonwebtoken
+#[cfg(feature = "use_pem")]
 #[test]
 fn roundtrip_with_jwtio_example() {
     // We currently do not support SEC1 so we use the converted PKCS8 formatted